requirements.txt
robotframework
robotframework-browser
robotframework-faker
robotframework-requests
Faker
mimesis
